move NoticInfo Interface
