+++ mod.ts
@@ -1,3 +1,2 @@
 export { getToken } from "./token.ts";
-export { NoticeInfo } from "./kupid.ts";
-export { KupidType, fetchKupidNotices, fetchParsedKupidNoptices } from "./wrapper.ts";
+export { KupidType, fetchKupidNotices, fetchParsedKupidNoptices, NoticeInfo } from "./wrapper.ts";
